2.18

+++ database/migrations/2017_08_08_115903_create_products_table.php
@@ -31,7 +31,7 @@ class CreateProductsTable extends Migration
             $table->foreign('tax')->references('id')->on('tax_rates');
             $table->enum('tax_type', ['inclusive', 'exclusive']);
             $table->boolean('enable_stock')->default(0);
-            $table->integer('alert_quantity')->default(0);
+            $table->decimal('alert_quantity', 22, 4)->default(0);
             $table->string('sku');
             $table->enum('barcode_type', ['C39', 'C128', 'EAN-13', 'EAN-8', 'UPC-A', 'UPC-E', 'ITF-14']);
             $table->integer('created_by')->unsigned();
